requirements.txt
Django==5.1.5
django-filter==24.3
djangorestframework==3.15.2
django-cors-headers==4.6.0
gunicorn==23.0.0
python-dotenv==1.0.1
psycopg2-binary==2.9.10
requests==2.32.3